% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterPhylip}
\alias{clusterPhylip}
\title{Cluster Description}
\usage{
clusterPhylip(
  phylip_path,
  cutoff,
  method,
  count_table,
  simularity_matrix = FALSE
)
}
\arguments{
\item{phylip_path}{A phylip file path.}

\item{cutoff}{A cutoff value.}

\item{method}{The type of cluster you wish to conduct;
furthest, nearest, average, weighted.}

\item{count_table}{A table of names and the given abundance per group.}

\item{simularity_matrix}{are you using a simularity matrix or
distance matrix.}
}
\value{
A string of the given cluster.
}
\description{
Detailed description of the function.
}
