% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clustur}
\alias{clustur}
\title{Clustur Description}
\usage{
clustur(
  sparse_matrix,
  cutoff,
  method,
  count_table,
  simularity_matrix = FALSE,
  random_seed = 123
)
}
\arguments{
\item{sparse_matrix}{The sparse matrix that you want to cluster with.}

\item{cutoff}{The value you wish to use as a cutoff when clustering.}

\item{method}{The type of cluster you wish to conduct;
opti, furthest, nearest, average, or weighted.}

\item{count_table}{A table of names and the given abundance per group.}

\item{simularity_matrix}{are you using a simularity matrix or distance matrix?}

\item{random_seed}{the random seed you wish to use, defaulted at 123.}
}
\value{
A string of the given cluster.
}
\description{
You must specfiy the type of matrix you are inputting
 to cluster your object and we support three types:
the path to your phylip and column distance file, or a sparse matrix.
}
\details{
Detailed description of the function.
}
\examples{

 # Convert Phylip of column file to a sparse matrix
 column_path <- example_path("updated_column.dist")
 phylip_path <- example_path("updated_phylip_1.txt")
 count_table <- readRDS(example_path("count_table.RDS"))

 phylip_sparse <- convert_distance_file_to_sparse(count_table, phylip_path, "phylip")
 column_sparse <- convert_distance_file_to_sparse(count_table, column_path, "column")
 cutoff <- 0.2
 # The clustur using one of the 5 methods
 # opti
 cluster_results <- clustur(column_sparse, cutoff, "opti", count_table)
 # furthest 
 cluster_results <- clustur(phylip_sparse, cutoff, "furthest", count_table)
 # nearest
 cluster_results <- clustur(column_sparse, cutoff, "nearest", count_table)
 # average
 cluster_results <- clustur(phylip_sparse, cutoff, "average", count_table)
 # weighted
 cluster_results <- clustur(column_sparse, cutoff, "weighted", count_table)



}
