% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{convert_distance_file_to_sparse}
\alias{convert_distance_file_to_sparse}
\title{Convert Distance File To Sparse}
\usage{
convert_distance_file_to_sparse(count_table, distance_file_path, type_of_file)
}
\arguments{
\item{count_table}{Your count table based on the data.}

\item{distance_file_path}{The path to the data file.}

\item{type_of_file}{What type of file you are passing to the function, has to be "phylip", or "column".}
}
\value{
a sparse matrix based off the data provided
}
\description{
Convert Distance File To Sparse
}
\examples{
# This will return the path to our example file
column_path <- example_path("updated_column.dist")
phylip_path <- example_path("updated_phylip_1.txt")
count_table <- readRDS(example_path("count_table.RDS"))

phylip_sparse <- convert_distance_file_to_sparse(count_table, phylip_path, "phylip")
column_sparse <- convert_distance_file_to_sparse(count_table, column_path, "column")
}
