% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{example_path}
\alias{example_path}
\title{Example Path}
\usage{
example_path(file = NULL)
}
\value{
the path inside of the package of the file.
}
\description{
Example Path
}
\examples{
# This will return the path to our example file
example_path("98_sq_phylip_amazon.dist")

}
