% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_object-getters.R
\name{get_abundance}
\alias{get_abundance}
\title{Get Shared}
\usage{
get_abundance(cluster_data)
}
\arguments{
\item{cluster_data}{The output from the `cluster()` function.}
}
\value{
a shared data.frame
}
\description{
GetShared returns the generated abundance `data.frame`
from the `cluster()` function
}
\examples{
cutoff <- 0.2
count_table <- read_count(example_path("amazon.full.count_table"))
distance_data <- read_dist(example_path("amazon_column.dist"),
                           count_table, cutoff, FALSE)
df_clusters <- cluster(distance_data, cutoff, method = "opticlust")
shared <- get_abundance(df_clusters)

}
