% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_object-getters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get Clusters}
\usage{
get_clusters(cluster_data)
}
\arguments{
\item{cluster_data}{The cluster method output}
}
\value{
the cluster data.frame
}
\description{
This function returns the cluster data.frame of the cluster object
}
\examples{
cutoff <- 0.2
count_table <- read_count(example_path("amazon.count_table"))
distance_data <- read_dist(example_path("amazon_column.dist"), count_table, cutoff, FALSE)
df_clusters <- cluster(distance_data, method = "opti")
clusters <- get_clusters(df_clusters)

}
