% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_object-getters.R
\name{get_count_table}
\alias{get_count_table}
\title{Get Count Table}
\usage{
get_count_table(distance_object)
}
\arguments{
\item{distance_object}{The read.dist output object}
}
\value{
a count table data.frame
}
\description{
This function returns the count table that was used
to generate the distance object.
}
\examples{
count_table <- read_count(example_path("amazon.count_table"))
distance_data <- read_dist(example_path("amazon_column.dist"), count_table, cutoff, FALSE)
count_table <- get_count_table(distance_data)

}
