% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_object-getters.R
\name{get_distance_df}
\alias{get_distance_df}
\title{Get Distance Data Frame}
\usage{
get_distance_df(distance_object)
}
\arguments{
\item{distance_object}{The output from the `read.dist()` function.}
}
\value{
a distance `data.frame`.
}
\description{
This function will generate a `data.frame` that contains the distances
of all the indexes.
}
\examples{
cutoff <- 0.2
count_table <- read_count(example_path("amazon.full.count_table"))
distance_data <- read_dist(example_path("amazon_column.dist"),
                                        count_table, cutoff, FALSE)
count_table <- get_count_table(distance_data)

}
