% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_object-getters.R
\name{get_shared}
\alias{get_shared}
\title{Get Shared}
\usage{
get_shared(cluster_data)
}
\arguments{
\item{cluster_data}{The cluster method output}
}
\value{
a shared data.frame
}
\description{
This function returns the shared data.frame of the cluster object
}
\examples{
count_table <- read_count(example_path("amazon.count_table"))
distance_data <- read_dist(example_path("amazon_column.dist"), count_table, cutoff, FALSE)
df_clusters <- cluster(distance_data, method = "opti")
shared <- get_shared(df_clusters)

}
