% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{opti_cluster}
\alias{opti_cluster}
\title{Opticluster Description}
\usage{
opti_cluster(sparse_matrix, cutoff, iterations)
}
\arguments{
\item{param1}{A Sparse Matrix.}

\item{param2}{A cutoff value}

\item{param3}{The number of iterations}
}
\value{
A dataframe of the clusters.
}
\description{
Detailed description of the function.
}
