% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count.R
\name{read_count}
\alias{read_count}
\alias{read_count.strollur}
\alias{read_count.character}
\title{Read count table}
\usage{
read_count(count_table)

\method{read_count}{strollur}(count_table)

\method{read_count}{character}(count_table)
}
\arguments{
\item{count_table}{The file path of your count table or strollur object.}
}
\value{
a count table `data.frame`.
}
\description{
This function will read and return your count table. It can take in
sparse and full count tables.
}
\examples{
count_table <- read_count(example_path("amazon.full.count_table"))
}
