% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{read_dist}
\alias{read_dist}
\title{Read Dist Description}
\usage{
read_dist(distance_file, count_table, cutoff, is_simularity_matrix)
}
\arguments{
\item{distance_file}{Either a phylip or column distance file, or a sparse matrix.}

\item{count_table}{A table of names and the given abundance per group.}

\item{cutoff}{The value you wish to use as a cutoff when clustering.
opti, furthest, nearest, average, or weighted.}

\item{is_simularity_matrix}{are you using a simularity matrix or distance matrix?}
}
\value{
A distance object that contains all your distance information
}
\description{
Detailed description of the function.
}
\examples{

 # Convert Phylip of column file to a sparse matrix

 i_values <- as.integer(1:100)
 j_values <- as.integer(sample(1:100, 100, TRUE))
 x_values <- as.numeric(runif(100, 0, 1))
 s_matrix <- Matrix::spMatrix(nrow=max(i_values),
                              ncol=max(i_values),
                              i=i_values,
                              j=j_values,
                              x=x_values)
 column_path <- example_path("96_sq_column_amazon.dist")
 phylip_path <- example_path("98_sq_phylip_amazon.dist")
 count_table <- readRDS(example_path("amazon1.count_table"))

data_column <- ProcessDistanceFiles(column_path, count_table, 0.2, F) 
data_phylip <- ProcessDistanceFiles(phylip_path, count_table, 0.2, F) 


}
