% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Description}
\usage{
cluster(distance_object, method, random_seed = 123)
}
\arguments{
\item{distance_object}{The distance object that was created using the read_dist function.}

\item{method}{The type of cluster you wish to conduct;
opti, furthest, nearest, average, or weighted.}

\item{random_seed}{the random seed you wish to use, defaulted at 123.}
}
\value{
A list of dataframes that contain abundance, and clustering results.
}
\description{
Detailed description of the function.
}
\examples{

 # Convert Phylip of column file to a sparse matrix
 column_path <- example_path("updated_column.dist")
 phylip_path <- example_path("updated_phylip_1.txt")
 count_table <- readRDS(example_path("count_table.RDS"))

 phylip_sparse <- convert_distance_file_to_sparse(count_table, phylip_path, "phylip")
 column_sparse <- convert_distance_file_to_sparse(count_table, column_path, "column")
 cutoff <- 0.2
 # The clustur using one of the 5 methods
 # opti
 cluster_results <- clustur(column_sparse, cutoff, "opti", count_table)
 # furthest 
 cluster_results <- clustur(phylip_sparse, cutoff, "furthest", count_table)
 # nearest
 cluster_results <- clustur(column_sparse, cutoff, "nearest", count_table)
 # average
 cluster_results <- clustur(phylip_sparse, cutoff, "average", count_table)
 # weighted
 cluster_results <- clustur(column_sparse, cutoff, "weighted", count_table)



}
