% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Description}
\usage{
cluster(cutoff, method, count_table, simularity_matrix = FALSE, ...)
}
\arguments{
\item{cutoff}{A cutoff value.}

\item{method}{The type of cluster you wish to conduct;
furthest, nearest, average, weighted.}

\item{count_table}{A table of names and the given abundance per group.}

\item{simularity_matrix}{are you using a simularity matrix or}

\item{sparse_matrix}{A Sparse Matrix.}

\item{pylip_path}{The path of your phylip file.}

\item{column_path}{The path to your column distance matrix file.
distance matrix.}
}
\value{
A string of the given cluster.
}
\description{
Detailed description of the function.
}
