% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Description}
\usage{
cluster(distance_object, method = "opti", random_seed = 123)
}
\arguments{
\item{distance_object}{The distance object that
was created using the `read_dist()` function.}

\item{method}{The type of cluster you wish to conduct;
opti, furthest, nearest, average, or weighted.}

\item{random_seed}{the random seed you wish to use, defaulted at 123.}
}
\value{
A list of `data.frames` that contain abundance, and clustering results.
If you used "opti", it will also return clustering metrics.
}
\description{
Detailed description of the function.
}
\examples{

 cutoff <- 0.2
 count_table <- read_count(example_path("amazon.count_table"))
 distance_data <- read_dist(example_path("amazon_column.dist"),
                            count_table, cutoff, FALSE)

 cluster_results <- cluster(distance_data, method = "opti")
 cluster_results <- cluster(distance_data, method = "furthest")
 cluster_results <- cluster(distance_data, method = "nearest")
 cluster_results <- cluster(distance_data, method = "average")
 cluster_results <- cluster(distance_data, method = "weighted")


}
