% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Opticluster Description}
\usage{
cluster(sparse_matrix, cutoff, method, count_table)
}
\arguments{
\item{sparse_matrix}{A Sparse Matrix.}

\item{cutoff}{A cutoff value}

\item{method}{The type of cluster you wish to conduct. There are four different types:
furthest, nearest, average, weighted.}

\item{count_table}{A table of abundances}
}
\value{
A string of the given cluster.
}
\description{
Detailed description of the function.
}
