% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Description}
\usage{
cluster(cutoff, method, count_table, simularity_matrix = FALSE, ...)
}
\arguments{
\item{cutoff}{A cutoff value.}

\item{method}{The type of cluster you wish to conduct;
furthest, nearest, average, weighted.}

\item{count_table}{A table of names and the given abundance per group.}

\item{simularity_matrix}{are you using a simularity matrix or distance matrix}

\item{...}{Either your phylip file or column file path, or a sparse distance matrix}
}
\value{
A string of the given cluster.
}
\description{
You must specfiy the type of matrix you are inputting to cluster your object and we support three types:
the path to your phylip and column distance file, or a sparse matrix.
}
\details{
Detailed description of the function.
}
