% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Description}
\usage{
cluster(distance_object, method, random_seed = 123)
}
\arguments{
\item{distance_object}{The distance object that was created using the read_dist function.}

\item{method}{The type of cluster you wish to conduct;
opti, furthest, nearest, average, or weighted.}

\item{random_seed}{the random seed you wish to use, defaulted at 123.}
}
\value{
A list of dataframes that contain abundance, and clustering results.
}
\description{
Detailed description of the function.
}
\examples{
 # Convert Phylip of column file to a sparse matrix
 cutoff <- 0.2
 count_table <- read_count(example_path("amazon.count_table"))
 distance_data <- read_dist(example_path("amazon_column.dist"), count_table, cutoff, FALSE)

 # The clustur using one of the 5 methods
 # opti
 cluster_results <- cluster(distance_data, method = "opti")
 # furthest 
 cluster_results <- cluster(distance_data, method = "furthest")
 # nearest
 cluster_results <- cluster(distance_data, method = "nearest")
 # average
 cluster_results <- cluster(distance_data, method = "average")
 # weighted
 cluster_results <- cluster(distance_data, method = "weighted")


}
