% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustur-package.R
\docType{package}
\name{clustur-package}
\alias{clustur}
\alias{clustur-package}
\title{clustur: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.schlosslab.org/Clustur/}
}

}
\keyword{internal}
