% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{example_path}
\alias{example_path}
\title{Example Path}
\usage{
example_path(file = NULL)
}
\arguments{
\item{file}{The file name of the data; leave as NULL (default) to get full
list of example files}
}
\value{
the path to the file as a `character` or a vector of `character`
giving example filenames if `fill = NULL`.
}
\description{
This function was created as a helper function to generate file paths to our
internal data. You should use this function if you
want to follow along with the example, or interact with the data
}
\examples{
example_path("amazon_phylip.dist")
example_path()
}
