% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{example_path}
\alias{example_path}
\title{Example Path}
\usage{
example_path(file = NULL)
}
\arguments{
\item{file}{The file name of the data}
}
\value{
the path to the file as a `chr`.
}
\description{
This function was created as a helper function to generate file paths to our
 internal data. You should use this function if you
 want to follow along with the example, or intereact with the data
}
\examples{
# This will return the path to our example file
example_path("amazon_phylip.dist")

}
