% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{example_path}
\alias{example_path}
\title{Example Path Description}
\usage{
example_path(file = NULL)
}
\value{
path
}
\description{
This function was created as a helper function to generate file paths to our internal data. You are able to access
this function if you want to follow along with the example.
}
