% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{example_path}
\alias{example_path}
\title{Example Path}
\usage{
example_path(file = NULL)
}
\arguments{
\item{file}{The data of the path you are looking to find.}
}
\value{
the path inside of the package of the file.
}
\description{
This function was created as a helper function to generate file paths to our
 internal data. You are able to access this function if you
 want to follow along with the example.
}
\examples{
# This will return the path to our example file
example_path("amazon_phylip.dist")

}
