% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{opti_cluster}
\alias{opti_cluster}
\title{Opticluster Description}
\usage{
opti_cluster(
  sparse_matrix,
  cutoff,
  iterations = 100,
  shuffle = TRUE,
  count_table
)
}
\arguments{
\item{sparse_matrix}{A Sparse Matrix.}

\item{cutoff}{A cutoff value}

\item{iterations}{The number of iterations}

\item{shuffle}{a boolean to determine whether or not you want to shuffle the data before you cluster}

\item{count_table}{A table of abundances}
}
\value{
A data.frame of the clusters.
}
\description{
Detailed description of the function.
}
