% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{opti_cluster}
\alias{opti_cluster}
\title{Opticluster Description}
\usage{
opti_cluster(
  cutoff,
  count_table,
  iterations = 100,
  shuffle = TRUE,
  simularity_matrix = FALSE,
  ...
)
}
\arguments{
\item{cutoff}{A cutoff value}

\item{count_table}{A table of names and the given abundance per group.}

\item{iterations}{The number of iterations}

\item{shuffle}{a boolean to determine whether or
not you want to shuffle the data before you cluster}

\item{simularity_matrix}{are you using a simularity matrix or distance matrix}

\item{sparse_matrix}{Your Sparse Matrix}

\item{pylip_path}{The path of your phylip file.}

\item{column_path}{The path to your column distance matrix file.}
}
\value{
A data.frame of the cluster and cluster metrics.
}
\description{
Detailed description of the function.
}
