% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{read_count}
\alias{read_count}
\title{Read count table}
\usage{
read_count(count_table_path)
}
\arguments{
\item{count_table_path}{The file path of your count table.}
}
\value{
a count table `data.frame`.
}
\description{
This function will read and return your count table. It can take in
sparse and full count tables.
}
\examples{
count_table <- read_count(example_path("amazon.full.count_table"))
}
