% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{read_dist}
\alias{read_dist}
\title{Read Dist Description}
\usage{
read_dist(distance_file, count_table, cutoff, is_simularity_matrix = FALSE)
}
\arguments{
\item{distance_file}{Either a phylip or column
distance file, or a sparse matrix.}

\item{count_table}{A table of names and the given abundance per group.}

\item{cutoff}{The value you wish to use as a cutoff when clustering.}

\item{is_simularity_matrix}{are you using a
simularity matrix or distance matrix?}
}
\value{
An distance `externalptr` that contains all your 
distance information.
}
\description{
Detailed description of the function.
}
\examples{

 # Convert Phylip or column file to a sparse matrix
 # spMatrix requires the use of the matrix library
 i_values <- as.integer(1:100)
 j_values <- as.integer(sample(1:100, 100, TRUE))
 x_values <- as.numeric(runif(100, 0, 1))
 s_matrix <- create_sparse_matrix(i_values, j_values, x_values)


 column_path <- example_path("amazon_column.dist")
 phylip_path <- example_path("amazon_phylip.dist")
 count_table <- read_count(example_path("amazon.count_table"))

data_column <- read_dist(column_path, count_table, 0.2, FALSE)
data_phylip <- read_dist(phylip_path, count_table, 0.2, FALSE)
data_sparse <- read_dist(s_matrix, count_table, 0.2, FALSE)


}
