% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{validate_count_table}
\alias{validate_count_table}
\title{Validate Count Table}
\usage{
validate_count_table(count_table_df)
}
\arguments{
\item{count_table_df}{The count table `data.frame` object.}
}
\value{
A validated count table `data.frame`
}
\description{
If the count table is already valid nothing will change,
otherwise it will add a new group to the count table file.
}
\details{
Determines whether user supplied count table is valid
}
\examples{
 count_table <- read.delim(example_path("amazon.full.count_table"))
 count_table_valid <- validate_count_table(count_table)

}
