% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{validate_count_table}
\alias{validate_count_table}
\title{Opticluster Description}
\usage{
validate_count_table(count_table_df)
}
\arguments{
\item{count_table_df}{The count table, which contains all your abundance information for each sequence.}
}
\value{
Validated count table.
}
\description{
Detailed description of the function.
}
